+++ Client/src/components/Home/Pagination.js
@@ -1,17 +1,16 @@
 import React from "react";
 
 const Pagination = ({
-  filteredProducts,
-  productsPerPage,
+  totalPages,
   currentPage,
   paginate,
 }) => {
-  const pageNumbers = Math.ceil(filteredProducts.length / productsPerPage);
+  if (totalPages <= 1) return null;
 
   return (
     <div className="flex justify-center w-full md:w-3/4 p-4">
       <ul className="flex">
-        {Array.from({ length: pageNumbers }).map((_, index) => (
+        {Array.from({ length: totalPages }).map((_, index) => (
           <li key={index}>
             <button
               className={`px-3 py-2 mx-1 rounded-lg bg-gray-900 text-white ${